Issue #1406: Fix markup bug affecting strinsert()

+++ src/markup.c
@@ -2511,7 +2511,7 @@ ansi_string_replace(ansi_string *dst, int loc, int count, ansi_string *src)
   /* Move markup as necessary. */
   if (dstleft > 0) {
     memmove(dst->markup + srcend, dst->markup + (loc + count),
-            dstleft * sizeof(int16_t));
+            dstleft * sizeof(int32_t));
   }
 
   /* If, and only if, mis and mie have a markup_information in common,
